# swagger.yaml
openapi: 3.0.0
info:
  title: Signed URL Generator API
  version: 1.0.0
  description: A service to generate signed URLs
servers:
  - url: http://localhost:3000
    description: Development server
paths:
  /:
    get:
      summary: Welcome message
      description: Returns a welcome message for the API
      responses:
        '200':
          description: Welcome message
          content:
            application/json:
              schema:
                type: object
                properties:
  /generate:
    post:
      summary: Generate a signed URL
      description: Generates a URL with signature based on username and message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - message
              properties:
                username:
                  type: string
                  description: Username for the signed URL
                message:
                  type: string
                  description: Message to include in the signed URL
      responses:
        '200':
          description: Successfully generated signed URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  signedUrl:
                    type: string
                    description: The generated signed URL
                  expiresAt:
                    type: string
                    format: date-time
                    description: Expiration date and time
                  cached:
                    type: boolean
                    description: Whether the URL was retrieved from cache
        '400':
          description: Bad request - missing required fields
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /verify:
    get:
      summary: Verify a signed URL
      description: Verifies the signature and expiration of a signed URL
      parameters:
        - in: query
          name: u
          schema:
            type: string
          required: true
          description: Username from the URL
        - in: query
          name: t
          schema:
            type: string
          required: true
          description: Timestamp from the URL
        - in: query
          name: sig
          schema:
            type: string
          required: true
          description: Signature from the URL
      responses:
        '200':
          description: Valid signed URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                  username:
                    type: string
                  expiresAt:
                    type: string
                    format: date-time
        '401':
          description: Invalid or expired URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /active-urls:
    get:
      summary: List active signed URLs
      description: Lists all active signed URLs in the cache
      responses:
        '200':
          description: List of active URLs
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                    description: Number of active URLs
                  urls:
                    type: array
                    items:
                      type: object
                      properties:
                        username:
                          type: string
                        message:
                          type: string
                        signedUrl:
                          type: string
                        expiresAt:
                          type: string
                          format: date-time
components: {}
tags: []